Add group centroid location to an existing plot

Add group centroid location to an existing plot

addCentroids(
  input,
  type,
  tag = NULL,
  track = NULL,
  timeslot = NULL,
  shape = 21,
  size = 1.5,
  colour = "white",
  fill = "cyan"
)

Arguments

input

The output of getCentroids

type

One of "group" or "track".

tag

Animal of interest, when type = "track".

track

Track of interest, when type = "track".

timeslot

The timeslot of interest to plot the centroid location

shape

The shape of the points

size

The size of the points

colour

The colour of the points

fill

The fill of the points

Value

A ggplot with centroid locations

Examples

# \donttest{
# Import river shapefile
water <- actel::loadShape(path = system.file(package = "RSP"), 
 shape = "River_latlon.shp", size = 0.0001, buffer = 0.05) 
#> Error in actel::loadShape(path = system.file(package = "RSP"), shape = "River_latlon.shp",     size = 1e-04, buffer = 0.05): unused argument (path = system.file(package = "RSP"))

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)
#> M: Constructing the transition layer. This process may take several minutes depending on the study area size and chosen pixel size.
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'transition': error in evaluating the argument 'x' in selecting a method for function 'raster': object 'water' not found

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")
#> M: Calculating RSP for the 'explore' data compiled on 2020-09-17 12:14:48.354614
#>   |                                                                    |                                                            |   0%
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'shortestPath': object 'tl' not found
#> Timing stopped at: 0.022 0.001 0.031

# Run dynamic Brownian Bridge Movement Model (dBBMM) with timeslots:
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56, timeframe = 2)
#> Error in eval(expr, envir, enclos): object 'water' not found

# Get dBBMM areas at group level
areas.group <- getAreas(dbbmm.data, type = "group", breaks = c(0.5, 0.95))
#> Error in eval(expr, envir, enclos): object 'dbbmm.data' not found

# Obtaing centroid coordinate locations of dBBMM:
df.centroid <- getCentroids(input = dbbmm.data, type = "group", areas = areas.group, 
   level = 0.95, group = "G1", UTM = 56)
#> Error in eval(expr, envir, enclos): object 'areas.group' not found

# Plot group centroid location:
plotAreas(areas.group, base.raster = water, group = "G1", timeslot = 7) +
   addCentroids(input = df.centroid, type = "group", timeslot = 7)
#> Error in eval(expr, envir, enclos): object 'areas.group' not found
# }