Animate the RSP tracks

This function can be used to generate an animated plot of the RSP tracks.

animateTracks(
  input,
  base.raster,
  tags = NULL,
  drop.groups = NULL,
  by.group = FALSE,
  start.time,
  stop.time,
  land.col = "#BABCBF80",
  add.legend = TRUE,
  add.stations = FALSE,
  save.gif = FALSE,
  gif.name = "Animation.gif",
  height = 720,
  width = 720,
  xlim = NULL,
  ylim = NULL,
  nframes = 100,
  fps = 10
)

Arguments

input

The output of runRSP.

base.raster

The raster used to generate the transition layer used in runRSP.

tags

Character vector specifying which tags to include in the animation.

drop.groups

Character vector specifying any group(s) to the be removed from the animation.

by.group

Logical, if TRUE one facet will be plotted for each tracked group. Defauly is FALSE.

start.time

Character vector of the start point (format = "Y-m-d H:M:S") for the animation.

stop.time

Character vector of the stop point (format = "Y-m-d H:M:S") for the animation.

land.col

Colour of the land masses. Defaults to semi-transparent grey.

add.legend

Logical, if TRUE (default) a colour legend representing the monitored tags will be included.

add.stations

Logical, if TRUE the stations will be added to the animaltion. Default is FALSE. Only works if by.group = FALSE.

save.gif

Logical defining if the animation should be saved.

gif.name

If save.gif = TRUE, character vector for the GIF name.

height

If save.gif = TRUE, number of pixels for the GIF height.

width

If save.gif = TRUE, number of pixels for the GIF width.

xlim

Numeric vector defining the horizontal limits of the map.

ylim

Numeric vector defining the vertical limits of the map.

nframes

The number of frames to render (default 100).

fps

The framerate of the animation in frames/sec (default 10).

Value

An animation of the RSP tracks.

Examples

# \donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 
#> M: Applying a small correction to the shape extent to ensure an integer number of pixels.
#> M: New shape extent:
#> SpatExtent : 150.948943357766, 151.109043357766, -33.9001889854332, -33.7532889854332 (xmin, xmax, ymin, ymax)
#> M: Chosen pixel size: 1e-04 
#> M: Resulting pixel dimensions:
#> xmax ymax 
#> 1601 1469 
#> M: Burning the shape into a raster. This process may take several minutes depending on the shape size and chosen pixel size.

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)
#> M: Constructing the transition layer. This process may take several minutes depending on the study area size and chosen pixel size.

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")
#> M: Calculating RSP for the 'explore' data compiled on 2020-09-17 12:14:48.354614
#>   |                                                                    |                                                            |   0%  |                                                                    |==============================                              |  50%  |                                                                    |============================================================| 100%
#> M: Percentage of detections valid for RSP: 100%

# Animate and RSP track:
animateTracks(input = rsp.data, base.raster = water, tags = "A69-9001-1111", add.stations = TRUE)
#> Preparing RSP data for the animation...
#>   [1] "./gganim_plot0001.png" "./gganim_plot0002.png" "./gganim_plot0003.png"
#>   [4] "./gganim_plot0004.png" "./gganim_plot0005.png" "./gganim_plot0006.png"
#>   [7] "./gganim_plot0007.png" "./gganim_plot0008.png" "./gganim_plot0009.png"
#>  [10] "./gganim_plot0010.png" "./gganim_plot0011.png" "./gganim_plot0012.png"
#>  [13] "./gganim_plot0013.png" "./gganim_plot0014.png" "./gganim_plot0015.png"
#>  [16] "./gganim_plot0016.png" "./gganim_plot0017.png" "./gganim_plot0018.png"
#>  [19] "./gganim_plot0019.png" "./gganim_plot0020.png" "./gganim_plot0021.png"
#>  [22] "./gganim_plot0022.png" "./gganim_plot0023.png" "./gganim_plot0024.png"
#>  [25] "./gganim_plot0025.png" "./gganim_plot0026.png" "./gganim_plot0027.png"
#>  [28] "./gganim_plot0028.png" "./gganim_plot0029.png" "./gganim_plot0030.png"
#>  [31] "./gganim_plot0031.png" "./gganim_plot0032.png" "./gganim_plot0033.png"
#>  [34] "./gganim_plot0034.png" "./gganim_plot0035.png" "./gganim_plot0036.png"
#>  [37] "./gganim_plot0037.png" "./gganim_plot0038.png" "./gganim_plot0039.png"
#>  [40] "./gganim_plot0040.png" "./gganim_plot0041.png" "./gganim_plot0042.png"
#>  [43] "./gganim_plot0043.png" "./gganim_plot0044.png" "./gganim_plot0045.png"
#>  [46] "./gganim_plot0046.png" "./gganim_plot0047.png" "./gganim_plot0048.png"
#>  [49] "./gganim_plot0049.png" "./gganim_plot0050.png" "./gganim_plot0051.png"
#>  [52] "./gganim_plot0052.png" "./gganim_plot0053.png" "./gganim_plot0054.png"
#>  [55] "./gganim_plot0055.png" "./gganim_plot0056.png" "./gganim_plot0057.png"
#>  [58] "./gganim_plot0058.png" "./gganim_plot0059.png" "./gganim_plot0060.png"
#>  [61] "./gganim_plot0061.png" "./gganim_plot0062.png" "./gganim_plot0063.png"
#>  [64] "./gganim_plot0064.png" "./gganim_plot0065.png" "./gganim_plot0066.png"
#>  [67] "./gganim_plot0067.png" "./gganim_plot0068.png" "./gganim_plot0069.png"
#>  [70] "./gganim_plot0070.png" "./gganim_plot0071.png" "./gganim_plot0072.png"
#>  [73] "./gganim_plot0073.png" "./gganim_plot0074.png" "./gganim_plot0075.png"
#>  [76] "./gganim_plot0076.png" "./gganim_plot0077.png" "./gganim_plot0078.png"
#>  [79] "./gganim_plot0079.png" "./gganim_plot0080.png" "./gganim_plot0081.png"
#>  [82] "./gganim_plot0082.png" "./gganim_plot0083.png" "./gganim_plot0084.png"
#>  [85] "./gganim_plot0085.png" "./gganim_plot0086.png" "./gganim_plot0087.png"
#>  [88] "./gganim_plot0088.png" "./gganim_plot0089.png" "./gganim_plot0090.png"
#>  [91] "./gganim_plot0091.png" "./gganim_plot0092.png" "./gganim_plot0093.png"
#>  [94] "./gganim_plot0094.png" "./gganim_plot0095.png" "./gganim_plot0096.png"
#>  [97] "./gganim_plot0097.png" "./gganim_plot0098.png" "./gganim_plot0099.png"
#> [100] "./gganim_plot0100.png"
#> attr(,"frame_vars")
#>     frame nframes progress          frame_time
#> 1       1     100     0.01 2020-01-05 00:00:00
#> 2       2     100     0.02 2020-01-05 00:01:51
#> 3       3     100     0.03 2020-01-05 00:03:42
#> 4       4     100     0.04 2020-01-05 00:05:33
#> 5       5     100     0.05 2020-01-05 00:07:24
#> 6       6     100     0.06 2020-01-05 00:09:15
#> 7       7     100     0.07 2020-01-05 00:11:06
#> 8       8     100     0.08 2020-01-05 00:12:57
#> 9       9     100     0.09 2020-01-05 00:14:48
#> 10     10     100     0.10 2020-01-05 00:16:39
#> 11     11     100     0.11 2020-01-05 00:18:30
#> 12     12     100     0.12 2020-01-05 00:20:21
#> 13     13     100     0.13 2020-01-05 00:22:12
#> 14     14     100     0.14 2020-01-05 00:24:03
#> 15     15     100     0.15 2020-01-05 00:25:54
#> 16     16     100     0.16 2020-01-05 00:27:45
#> 17     17     100     0.17 2020-01-05 00:29:36
#> 18     18     100     0.18 2020-01-05 00:31:27
#> 19     19     100     0.19 2020-01-05 00:33:18
#> 20     20     100     0.20 2020-01-05 00:35:09
#> 21     21     100     0.21 2020-01-05 00:37:00
#> 22     22     100     0.22 2020-01-05 00:38:51
#> 23     23     100     0.23 2020-01-05 00:40:42
#> 24     24     100     0.24 2020-01-05 00:42:33
#> 25     25     100     0.25 2020-01-05 00:44:24
#> 26     26     100     0.26 2020-01-05 00:46:15
#> 27     27     100     0.27 2020-01-05 00:48:06
#> 28     28     100     0.28 2020-01-05 00:49:57
#> 29     29     100     0.29 2020-01-05 00:51:48
#> 30     30     100     0.30 2020-01-05 00:53:39
#> 31     31     100     0.31 2020-01-05 00:55:30
#> 32     32     100     0.32 2020-01-05 00:57:21
#> 33     33     100     0.33 2020-01-05 00:59:12
#> 34     34     100     0.34 2020-01-05 01:01:04
#> 35     35     100     0.35 2020-01-05 01:02:55
#> 36     36     100     0.36 2020-01-05 01:04:46
#> 37     37     100     0.37 2020-01-05 01:06:37
#> 38     38     100     0.38 2020-01-05 01:08:28
#> 39     39     100     0.39 2020-01-05 01:10:19
#> 40     40     100     0.40 2020-01-05 01:12:10
#> 41     41     100     0.41 2020-01-05 01:14:01
#> 42     42     100     0.42 2020-01-05 01:15:52
#> 43     43     100     0.43 2020-01-05 01:17:43
#> 44     44     100     0.44 2020-01-05 01:19:34
#> 45     45     100     0.45 2020-01-05 01:21:25
#> 46     46     100     0.46 2020-01-05 01:23:16
#> 47     47     100     0.47 2020-01-05 01:25:07
#> 48     48     100     0.48 2020-01-05 01:26:58
#> 49     49     100     0.49 2020-01-05 01:28:49
#> 50     50     100     0.50 2020-01-05 01:30:40
#> 51     51     100     0.51 2020-01-05 01:32:31
#> 52     52     100     0.52 2020-01-05 01:34:22
#> 53     53     100     0.53 2020-01-05 01:36:13
#> 54     54     100     0.54 2020-01-05 01:38:04
#> 55     55     100     0.55 2020-01-05 01:39:55
#> 56     56     100     0.56 2020-01-05 01:41:46
#> 57     57     100     0.57 2020-01-05 01:43:37
#> 58     58     100     0.58 2020-01-05 01:45:28
#> 59     59     100     0.59 2020-01-05 01:47:19
#> 60     60     100     0.60 2020-01-05 01:49:10
#> 61     61     100     0.61 2020-01-05 01:51:01
#> 62     62     100     0.62 2020-01-05 01:52:52
#> 63     63     100     0.63 2020-01-05 01:54:43
#> 64     64     100     0.64 2020-01-05 01:56:34
#> 65     65     100     0.65 2020-01-05 01:58:25
#> 66     66     100     0.66 2020-01-05 02:00:16
#> 67     67     100     0.67 2020-01-05 02:02:08
#> 68     68     100     0.68 2020-01-05 02:03:59
#> 69     69     100     0.69 2020-01-05 02:05:50
#> 70     70     100     0.70 2020-01-05 02:07:41
#> 71     71     100     0.71 2020-01-05 02:09:32
#> 72     72     100     0.72 2020-01-05 02:11:23
#> 73     73     100     0.73 2020-01-05 02:13:14
#> 74     74     100     0.74 2020-01-05 02:15:05
#> 75     75     100     0.75 2020-01-05 02:16:56
#> 76     76     100     0.76 2020-01-05 02:18:47
#> 77     77     100     0.77 2020-01-05 02:20:38
#> 78     78     100     0.78 2020-01-05 02:22:29
#> 79     79     100     0.79 2020-01-05 02:24:20
#> 80     80     100     0.80 2020-01-05 02:26:11
#> 81     81     100     0.81 2020-01-05 02:28:02
#> 82     82     100     0.82 2020-01-05 02:29:53
#> 83     83     100     0.83 2020-01-05 02:31:44
#> 84     84     100     0.84 2020-01-05 02:33:35
#> 85     85     100     0.85 2020-01-05 02:35:26
#> 86     86     100     0.86 2020-01-05 02:37:17
#> 87     87     100     0.87 2020-01-05 02:39:08
#> 88     88     100     0.88 2020-01-05 02:40:59
#> 89     89     100     0.89 2020-01-05 02:42:50
#> 90     90     100     0.90 2020-01-05 02:44:41
#> 91     91     100     0.91 2020-01-05 02:46:32
#> 92     92     100     0.92 2020-01-05 02:48:23
#> 93     93     100     0.93 2020-01-05 02:50:14
#> 94     94     100     0.94 2020-01-05 02:52:05
#> 95     95     100     0.95 2020-01-05 02:53:56
#> 96     96     100     0.96 2020-01-05 02:55:47
#> 97     97     100     0.97 2020-01-05 02:57:38
#> 98     98     100     0.98 2020-01-05 02:59:29
#> 99     99     100     0.99 2020-01-05 03:01:20
#> 100   100     100     1.00 2020-01-05 03:03:12
#>                                         frame_source
#> 1   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0001.png
#> 2   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0002.png
#> 3   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0003.png
#> 4   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0004.png
#> 5   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0005.png
#> 6   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0006.png
#> 7   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0007.png
#> 8   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0008.png
#> 9   /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0009.png
#> 10  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0010.png
#> 11  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0011.png
#> 12  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0012.png
#> 13  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0013.png
#> 14  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0014.png
#> 15  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0015.png
#> 16  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0016.png
#> 17  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0017.png
#> 18  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0018.png
#> 19  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0019.png
#> 20  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0020.png
#> 21  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0021.png
#> 22  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0022.png
#> 23  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0023.png
#> 24  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0024.png
#> 25  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0025.png
#> 26  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0026.png
#> 27  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0027.png
#> 28  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0028.png
#> 29  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0029.png
#> 30  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0030.png
#> 31  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0031.png
#> 32  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0032.png
#> 33  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0033.png
#> 34  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0034.png
#> 35  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0035.png
#> 36  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0036.png
#> 37  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0037.png
#> 38  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0038.png
#> 39  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0039.png
#> 40  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0040.png
#> 41  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0041.png
#> 42  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0042.png
#> 43  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0043.png
#> 44  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0044.png
#> 45  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0045.png
#> 46  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0046.png
#> 47  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0047.png
#> 48  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0048.png
#> 49  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0049.png
#> 50  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0050.png
#> 51  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0051.png
#> 52  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0052.png
#> 53  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0053.png
#> 54  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0054.png
#> 55  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0055.png
#> 56  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0056.png
#> 57  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0057.png
#> 58  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0058.png
#> 59  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0059.png
#> 60  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0060.png
#> 61  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0061.png
#> 62  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0062.png
#> 63  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0063.png
#> 64  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0064.png
#> 65  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0065.png
#> 66  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0066.png
#> 67  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0067.png
#> 68  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0068.png
#> 69  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0069.png
#> 70  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0070.png
#> 71  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0071.png
#> 72  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0072.png
#> 73  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0073.png
#> 74  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0074.png
#> 75  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0075.png
#> 76  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0076.png
#> 77  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0077.png
#> 78  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0078.png
#> 79  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0079.png
#> 80  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0080.png
#> 81  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0081.png
#> 82  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0082.png
#> 83  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0083.png
#> 84  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0084.png
#> 85  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0085.png
#> 86  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0086.png
#> 87  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0087.png
#> 88  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0088.png
#> 89  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0089.png
#> 90  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0090.png
#> 91  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0091.png
#> 92  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0092.png
#> 93  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0093.png
#> 94  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0094.png
#> 95  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0095.png
#> 96  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0096.png
#> 97  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0097.png
#> 98  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0098.png
#> 99  /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0099.png
#> 100 /tmp/Rtmp8Dh1fT/25514282c4fe/gganim_plot0100.png
# }