Calculate overlaps between different groups

Calculate overlaps between different groups

getOverlaps(input)

Arguments

input

The output of getAreas

Value

A list of Overlaps (per timeslot if relevant), as well as the respective overlap rasters.

Examples

# \donttest{
# Import river shapefile
water <- actel::shapeToRaster(shape = paste0(system.file(package = "RSP"), "/River_latlon.shp"), 
size = 0.0001, buffer = 0.05) 
#> M: Applying a small correction to the shape extent to ensure an integer number of pixels.
#> M: New shape extent:
#> SpatExtent : 150.948943357766, 151.109043357766, -33.9001889854332, -33.7532889854332 (xmin, xmax, ymin, ymax)
#> M: Chosen pixel size: 1e-04 
#> M: Resulting pixel dimensions:
#> xmax ymax 
#> 1601 1469 
#> M: Burning the shape into a raster. This process may take several minutes depending on the shape size and chosen pixel size.

# Create a transition layer with 8 directions
tl <- actel::transitionLayer(x = water, directions = 8)
#> M: Constructing the transition layer. This process may take several minutes depending on the study area size and chosen pixel size.

# Import example output from actel::explore() 
data(input.example) 

# Run RSP analysis
rsp.data <- runRSP(input = input.example, t.layer = tl, coord.x = "Longitude", coord.y = "Latitude")
#> M: Calculating RSP for the 'explore' data compiled on 2020-09-17 12:14:48.354614
#>   |                                                                    |                                                            |   0%  |                                                                    |==============================                              |  50%  |                                                                    |============================================================| 100%
#> M: Percentage of detections valid for RSP: 100%

# Run dynamic Brownian Bridge Movement Model (dBBMM)
dbbmm.data <- dynBBMM(input = rsp.data, base.raster = water, UTM = 56)
#> M: Converting coordinates to UTM. Original latitude/longitude values for the detections will be stored in columns 'O.LAT' and 'O.LON'.
#> M: Preparing data to apply dBBMM.
#> M: No specific transmitters selected. All the data will be used for analysis.
#> Warning: 1 individual detections were removed in group G2 due to simultaneous detections at two receivers.
#> M: In total, 1 detections were excluded as they failed the track quality checks.
#> M: Calculating dBBMM: G1
#> M: Success! (Time spent: 00:00:05)
#> M: Calculating dBBMM: G2
#> M: Success! (Time spent: 00:00:05)
#> M: Subtracting land areas from output.

# Get dBBMM areas at group level
areas.group <- getAreas(dbbmm.data, type = "group", breaks = c(0.5, 0.95))

# Get overlaps between groups
overlap.data <- getOverlaps(areas.group)
#>   |                                                                    |                                                            |   0%  |                                                                    |==============================                              |  50%  |                                                                    |============================================================| 100%
# }